requirements.txt
albumentations==1.1.0
colorama==0.4.4
cycler==0.11.0
fonttools==4.29.1
imageio==2.16.0
imgaug==0.4.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
scikit-image==0.18.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Shapely==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.2.9
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
typing_extensions==4.1.1
